Parameterize remaining scheme functions
